removed unused types
